requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
python-multipart==0.0.20

tensorflow==2.16.1
keras==3.3.3
numpy==1.26.4
h5py==3.10.0

ml-dtypes==0.3.2
opt-einsum==3.4.0
absl-py==2.3.1
grpcio==1.75.1
six==1.17.0

pillow==11.0.0
opencv-python==4.11.0.86
matplotlib==3.10.6

mediapipe==0.10.21
protobuf>=4.25.3,<5
